makefile: fix lib/transform docs not getting updated
As of
https://github.com/rclone/rclone/commit/37c20e22d9faded7496accb295db2bbe2156f3c7
the lib/transform docs are generated with //go:generate and embedded with
//go:embed.

Before this change, however, they were not getting automatically updated with
subsequent changes (like
https://github.com/rclone/rclone/commit/464120d3be79886d82c0807962e21a848d4e3c62)
because `go generate ./lib/transform` was not being run as part of the release
making process.

This change fixes that by running it in `make commanddocs`.